success to pass data to the native function

+++ lib/AudioAnalyzer.dart
@@ -3,7 +3,8 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'dart:math';
 import 'dart:core';
-import 'dart:ffi'; // For FFI
+import 'dart:ffi' as ffi; // For FFI
+import 'package:ffi/ffi.dart' as ffi; // For free() that is not in the dart package
 import 'dart:io'; // For Platform.isX
 
 import 'package:flutter_audio_recorder/flutter_audio_recorder.dart';
@@ -15,23 +16,35 @@ import 'package:fft/fft.dart';
 //import 'package:smart_signal_processing/smart_signal_processing.dart' as ssp;
 import 'package:my_complex/my_complex.dart';
 
+typedef NativeFFTFunction = Function(ffi.Pointer<ffi.Int16>, int);
+
 class AudioAnalyzer {
 
     var _recorder;
     var _file;
     var _assetsAudioPlayer;
+    int Function(ffi.Pointer<ffi.Int16>, int) _libfftw_plugin_transform;
+    ffi.Pointer<ffi.Int16> _data;
 
     AudioAnalyzer() {
         // https://flutter.dev/docs/development/platform-integration/c-interop
-        final DynamicLibrary nativeAddLib = Platform.isAndroid
-            ? DynamicLibrary.open("libfftw_plugin.so")
-            : DynamicLibrary.process();
-        final int Function() transform =
-        nativeAddLib
-            .lookup<NativeFunction<Int32 Function()>>("transform")
+        // https://stackoverflow.com/questions/58838193/pass-uint8list-to-pointervoid-in-dartffi
+        // https://github.com/martin-labanic/camera_preview_ffi_image_processing/blob/master/image_processing_plugin/lib/image_processing_plugin.dart
+        final ffi.DynamicLibrary nativeAddLib = Platform.isAndroid
+            ? ffi.DynamicLibrary.open("libfftw_plugin.so")
+            : ffi.DynamicLibrary.process();
+
+        _libfftw_plugin_transform = nativeAddLib
+            .lookup<ffi.NativeFunction<
+            ffi.Int32 Function(ffi.Pointer<ffi.Int16>, ffi.Int32)
+            >>("transform")
             .asFunction();
-        var temp = transform();
-        var temp2 = 1;
+    }
+
+    void dispose() {
+        if (_data != null) {
+            ffi.free(_data);
+        }
     }
 
     makeFile() async {
@@ -53,7 +66,10 @@ class AudioAnalyzer {
         }
 
         var file = File(_file);
-        file.delete();
+        var doesFileExist = await file.exists();
+        if (doesFileExist) {
+            file.delete();
+        }
         // https://pub.dev/packages/flutter_audio_recorder
         // https://developer.android.com/ndk/guides/audio/sampling-audio
         // 48 can be used for high notes.
@@ -101,7 +117,7 @@ class AudioAnalyzer {
     }
 
     // In the README, we have an adb command to download the wav
-    // the we can parse the header in python
+    // then we can parse the header in python
     // python wav parser
     // https://gist.github.com/eerwitt/ba51e181d50de6555a2ae613a558c0b6
     // the WAV
@@ -110,14 +126,21 @@ class AudioAnalyzer {
 
     // windowing makes a more periodic signal
     // https://download.ni.com/evaluation/pxi/Understanding%20FFTs%20and%20Windowing.pdf
+
+    // source for the data setup to the native function:
+    // https://stackoverflow.com/questions/58838193/pass-uint8list-to-pointervoid-in-dartffi
     computeFFT() async {
-        var periodicSamples = await getPeriodicSamples();
-        var fft = new FFT().Transform(periodicSamples);
-        return fft;
+        Int16List periodicSamples = await getPeriodicSamplesSkippingTheHeader();
+        _data = ffi.allocate<ffi.Int16>(count: periodicSamples.length); // Allocate a pointer large enough.
+        final pointerList = _data.asTypedList(periodicSamples.length); // Create a list that uses our pointer and copy in the image data.
+        pointerList.setAll(0, periodicSamples);
+        var fft = _libfftw_plugin_transform(_data, periodicSamples.length);
+        var temp1 = periodicSamples[0];
+        return new List<double>();
     }
 
-    getPeriodicSamples() async {
-        var samples = await getSamples();
+    getPeriodicSamplesSkippingTheHeader() async {
+        Int16List samples = await getSamples();
 
         // we force the signal to be periodic by cutting at a certain mean amplitude value
         // because the sound is correct to the ear, the FFT must give good coefficients
@@ -125,8 +148,11 @@ class AudioAnalyzer {
         double mean1 = 0;
         double mean2 = 0;
         num tenPercentSize = samples.length / 10;
-        for (var i = 0; i < tenPercentSize; i++) {
+        // we start at 22 because we skip the 44 bytes of the wav file header
+        for (var i = 22; i < tenPercentSize; i++) {
             mean1 += samples[i];
+        }
+        for (var i = 0; i < tenPercentSize; i++) {
             mean2 += samples[samples.length - 1 - i];
         }
         mean1 /= tenPercentSize;
@@ -136,7 +162,7 @@ class AudioAnalyzer {
         int index2 = getIndexForCutting(samples, sideValue, true);
         samples[index1] = sideValue;
         samples[index2] = sideValue;
-        List<int> periodicSamples = samples.sublist(index1, index2 + 1);
+        Int16List periodicSamples = samples.sublist(index1, index2 + 1);
         return periodicSamples;
     }
 
@@ -144,29 +170,33 @@ class AudioAnalyzer {
         // the header has 44 bytes, which makes 22 16-bit integers
         // we know each byte has 2 bytes
         Uint8List bytes = await new File(_file).readAsBytes();
-        int lengthBytes = bytes.length - 44;
+        //int lengthBytes = bytes.length - 44;
         // the Hann function wants a sample with size equal to a power of 2
         // https://stackoverflow.com/questions/466204/rounding-up-to-next-power-of-2/24844826
-        var powerOf2 = pow(2, (log(lengthBytes) / log(2)).ceil()) / 2;
-        int extra = (lengthBytes - powerOf2) ~/ 2;
+        //var powerOf2 = pow(2, (log(lengthBytes) / log(2)).ceil()) / 2;
+        //int extra = (lengthBytes - powerOf2) ~/ 2;
 
         ByteBuffer buffer = bytes.buffer;
         // 16-bit samples are stored as 2's-complement signed integers, ranging from -32768 to 32767.
         // http://soundfile.sapp.org/doc/WaveFormat/
         // 22 is removed for the header
         // extra is removed to have a power of 2.
-        var samplesOn2Bytes = buffer.asInt16List().skip(22);
-        List<int> samples = new List<int>.from(samplesOn2Bytes);
-        return samples;
+        var samplesOn2Bytes = buffer.asInt16List();
+        return samplesOn2Bytes;
     }
 
-    getIndexForCutting(List<int> samples, int sideValue, bool rightSide) {
+    getIndexForCutting(Int16List samples, int sideValue, bool rightSide) {
         int tenPercentSize = samples.length ~/ 10;
         // https://stackoverflow.com/questions/50429660/is-there-a-constant-for-max-min-int-double-value-in-dart
-        const int int64MaxValue = 9223372036854775807;
-        int distance = int64MaxValue;
+        int distance = samples[0];
         int index = 0;
         for (var i = 0; i < tenPercentSize; i++) {
+            if (!rightSide) {
+                // we discare the byes for the wav file header
+                if (i < 22) {
+                    i = 22;
+                }
+            }
             var sampleValue = samples[i];
             if (rightSide) {
                 sampleValue = samples[samples.length - 1 - i];
